#817 disallow invalid dates; move toDate to utils file
