+++ package.json
@@ -51,6 +51,7 @@
         "swagger-ui-express": "^4.3.0",
         "uuid": "^8.3.2",
         "uuid-apikey": "^1.5.1",
+        "validate-date": "^2.0.0",
         "validator": ">=13.7.0",
         "winston": "^3.2.1",
         "yamljs": "^0.3.0"
