+++ src/controller/cve-id.controller/cve-id.middleware.js
@@ -15,25 +15,6 @@ function parsePostParams (req, res, next) {
   next()
 }
 
-// Sanitizer for dates
-function toDate (val) {
-  val = val.toUpperCase()
-  let value = val.match(/^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(\.\d+)?(|Z|((-|\+|\s)\d{2}:\d{2}))$/)
-  let result
-  if (value) {
-    value[0] = value[0].replace(' ', '+') // Re-add literal '+' which was stripped
-    result = new Date(value[0])
-  } else {
-    value = val.match(/^\d{4}-\d{2}-\d{2}$/)
-    if (value) {
-      result = new Date(`${value[0]}T00:00:00.000+00:00`)
-    } else {
-      result = null
-    }
-  }
-  return result
-}
-
 function parseError (req, res, next) {
   const err = validationResult(req).formatWith(({ location, msg, param, value, nestedErrors }) => {
     return { msg: msg, param: param, location: location }
@@ -47,6 +28,5 @@ function parseError (req, res, next) {
 module.exports = {
   parseGetParams,
   parsePostParams,
-  parseError,
-  toDate
+  parseError
 }
