+++ src/controller/cve-id.controller/index.js
@@ -4,8 +4,9 @@ const mw = require('../../middleware/middleware')
 const errorMsgs = require('../../middleware/errorMessages')
 const controller = require('./cve-id.controller')
 const { param, query } = require('express-validator')
-const { parseGetParams, parsePostParams, parseError, toDate } = require('./cve-id.middleware')
+const { parseGetParams, parsePostParams, parseError } = require('./cve-id.middleware')
 const CONSTANTS = require('../../constants')
+const toDate = require('../../utils/utils').toDate
 
 const CHOICES = [CONSTANTS.CVE_STATES.REJECTED, CONSTANTS.CVE_STATES.PUBLISHED, CONSTANTS.CVE_STATES.RESERVED]
 const MODIFYTARGETS = [CONSTANTS.CVE_STATES.REJECTED, CONSTANTS.CVE_STATES.RESERVED]
