+++ src/utils/utils.js
@@ -1,6 +1,7 @@
 const Org = require('../model/org')
 const User = require('../model/user')
 const CONSTANTS = require('../constants')
+const validateDate = require('validate-date')
 
 async function getOrgUUID (shortName) {
   const org = await Org.findOne().byShortName(shortName)
@@ -100,6 +101,30 @@ function reqCtxMapping (req, keyType, keys) {
   }
 }
 
+// Sanitizer for dates
+function toDate (val) {
+  val = val.toUpperCase()
+  let value = val.match(/^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(\.\d+)?(|Z|((-|\+|\s)\d{2}:\d{2}))$/)
+  let result = null
+  if (value) {
+    value[0] = value[0].replace(' ', '+') // Re-add literal '+' which was stripped
+    const dateStr = value[0]
+    // Make sure that the string passed is a valid date
+    // eslint doesn't like that responseType is not defined, but it is needed as is
+    /* eslint-disable-next-line */
+    const valid = validateDate(dateStr.toString().substring(0, 10), responseType = 'boolean')
+    if (valid) {
+      result = new Date(dateStr)
+    }
+  } else {
+    value = val.match(/^\d{4}-\d{2}-\d{2}$/)
+    if (value) {
+      result = new Date(`${value[0]}T00:00:00.000+00:00`)
+    }
+  }
+  return result
+}
+
 module.exports = {
   isSecretariat,
   isAdmin,
@@ -107,5 +132,6 @@ module.exports = {
   isSecretariatUUID,
   getOrgUUID,
   getUserUUID,
-  reqCtxMapping
+  reqCtxMapping,
+  toDate
 }
